+++ app/Http/Controllers/Admin/SchoolYearController.php
@@ -2,20 +2,16 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Models\Honor;
 use App\Models\SchoolYear;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreHonorRequest;
-use App\Http\Requests\UpdateHonorRequest;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreSchoolYearRequest;
 
 class SchoolYearController extends Controller
 {
-
-    protected $menuHeader = "System Settings";
-    protected $menu = "system-settings";
 
     /**
      * Display a listing of the resource.
@@ -24,22 +20,11 @@ class SchoolYearController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('school_year_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-
-        $menuHeader = $this->menuHeader;
-        $menu = $this->menu;
-
-
-        $title = "School Years";
-        $breadcrumb = [
-            [
-                "name" => "School Years"
-            ]
-        ];
+        abort_if(Gate::denies('school_year_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $schoolYears = SchoolYear::order()->paginate(10);
 
-        return view('admin.school_years.index', compact('menuHeader', 'title', 'menu', 'breadcrumb', 'schoolYears'));
+        return view('admin.school_years.index', compact('schoolYears'));
     }
 
     /**
@@ -49,24 +34,9 @@ class SchoolYearController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('school_year_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('school_year_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $menuHeader = $this->menuHeader;
-        $menu = $this->menu;
-
-
-        $title = "Add School Year";
-        $breadcrumb = [
-            [
-                "name" => "School Years",
-                "url" => route('settings.school-years.index')
-            ],
-            [
-                "name" => "Add School Year"
-            ]
-        ];
-
-        return view('admin.school_years.create', compact('menuHeader', 'title', 'menu', 'breadcrumb'));
+        return view('admin.school_years.create');
     }
 
     /**
@@ -91,7 +61,7 @@ class SchoolYearController extends Controller
      */
     public function destroy(SchoolYear $schoolYear)
     {
-        // abort_if(Gate::denies('school_year_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('school_year_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $schoolYear->delete();
